requirements.txt
psycopg2
Pillow
wagtail
django-storages
easy-thumbnails
boto3
django-extensions
gunicorn
